issue-#64 全課題試み

+++ app/controllers/assigns_controller.rb
@@ -14,17 +14,8 @@ class AssignsController < ApplicationController
 
   def destroy
     assign = Assign.find(params[:id])
-
-    binding.irb
     destroy_message = assign_destroy(assign, assign.user)
-
-    # if current_user.id == assign.user.id || assign.team.owner
-    #   assign.destroy
-
-      redirect_to team_url(params[:team_id]), notice: destroy_message
-    # else
-    #   render team_url(params[:team_id]), notice:
-    # end
+    redirect_to team_url(params[:team_id]), notice: destroy_message
   end
 
   private
@@ -34,17 +25,19 @@ class AssignsController < ApplicationController
   end
 
   def assign_destroy(assign, assigned_user)
-    if assigned_user == assign.team.owner
-      I18n.t('views.messages.cannot_delete_the_leader')
-
-    binding.irb
-    elsif Assign.where(user_id: assigned_user.id).count == 1
-      I18n.t('views.messages.cannot_delete_only_a_member')
-    elsif assign.destroy
-      set_next_team(assign, assigned_user)
-      I18n.t('views.messages.delete_member')
-    else
-      I18n.t('views.messages.cannot_delete_member_4_some_reason')
+    if assign.team.owner == current_user || assign.user == current_user
+      if assigned_user == assign.team.owner
+        I18n.t('views.messages.cannot_delete_the_leader')
+      elsif Assign.where(user_id: assigned_user.id).count == 1
+        I18n.t('views.messages.cannot_delete_only_a_member')
+      elsif assign.destroy
+        set_next_team(assign, assigned_user)
+        I18n.t('views.messages.delete_member')
+      else
+        I18n.t('views.messages.cannot_delete_member_4_some_reason')
+      end
+    else 
+      "You are not authorized to perform such task"
     end
   end
